+++ src-tauri/src/commands/keybinds/get_keybinds.rs
@@ -1,9 +1,10 @@
 use tauri::{ AppHandle, State };
+
 use crate::{
+    ActiveProfile,
     config::keybinds::get_default_keybinds,
     models::keybinds::KeybindRegistry,
     utils::paths::get_profile_dir,
-    ActiveProfile,
 };
 
 #[tauri::command]
@@ -11,7 +12,7 @@ pub async fn get_keybinds(
     app: AppHandle,
     active_profile: State<'_, ActiveProfile>
 ) -> Result<KeybindRegistry, String> {
-    let defaults = get_default_keybinds();
+    let mut defaults = get_default_keybinds();
 
     let profile_dir = get_profile_dir(&app, &active_profile);
     let path = profile_dir.join("keybinds.json");
@@ -21,14 +22,19 @@ pub async fn get_keybinds(
     }
 
     let file_data = std::fs::read_to_string(&path).map_err(|e| e.to_string())?;
+    let saved_registry: KeybindRegistry = match serde_json::from_str(&file_data) {
+        Ok(data) => data,
+        Err(_) => {
+            eprintln!("Old keybind format detected. Resetting to defaults.");
+            return Ok(defaults);
+        }
+    };
 
-    let mut user_registry: KeybindRegistry = serde_json
-        ::from_str(&file_data)
-        .map_err(|e| e.to_string())?;
-
-    for (key, def) in defaults {
-        user_registry.entry(key).or_insert(def);
+    for default_item in defaults.iter_mut() {
+        if let Some(saved_item) = saved_registry.iter().find(|s| s.id == default_item.id) {
+            default_item.config = saved_item.config.clone();
+        }
     }
 
-    Ok(user_registry)
+    Ok(defaults)
 }