+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be logged here
 
 ## 1.4.0
 
+### Features
+
+- ✨ add Per profile keybinds
+
+### Bug Fixes
+
+- 🛠️ fix keybind ordering
+
 ### Dev Notes
 
 - ✏️ reduce boilerplate code with `Card.tsx` and `ListItem.tsx`